+++ crates/cubek-tile/src/tile/procedural/affine.rs
@@ -1,0 +1,24 @@
+use cubecl::prelude::*;
+
+use crate::{Axis, Coords, Space};
+
+use super::{Recipe, RecipeExpand};
+
+/// A one-dimensional affine coordinate expression, `offset + coefficient * coordinate[axis]`.
+/// The axis is compile-time metadata; offset and coefficient can be runtime values.
+#[derive(CubeType, Clone)]
+#[expand(derive(Clone))]
+pub struct AffineCoordinates<T: Numeric> {
+    pub offset: T,
+    pub coefficient: T,
+    #[cube(comptime)]
+    pub axis: Axis,
+}
+
+#[cube]
+impl<T: Numeric> Recipe<T> for AffineCoordinates<T> {
+    fn evaluate(&self, coordinates: &Coords<u32>, #[comptime] space: Space) -> T {
+        self.offset
+            + self.coefficient * T::cast_from(coordinates.at(comptime!(space.position(self.axis))))
+    }
+}
