+++ crates/cubek-tile/src/tile/procedural/mod.rs
@@ -1,0 +1,13 @@
+//! A memory-free tile source evaluated from logical coordinates.
+
+mod affine;
+mod base;
+mod constant;
+pub mod filter;
+mod recipe;
+
+pub use affine::*;
+pub use base::*;
+pub use constant::*;
+pub use filter::*;
+pub use recipe::*;
